// src/components/model/model.ts
export interface Employee {
  employeeID: string;
  firstName: string;
  lastName: string;
  company: string;
  phoneNo: string;
  email: string;

  address1: string;
  address2: string;
  address3: string;
  address4: string;

  skill1: string;
  skill1Experience: string;
  skill2: string;
  skill2Experience: string;
  skill3: string;
  skill3Experience: string;
  skill4: string;
  skill4Experience: string;

  fathersName: string;
  fathersProfession: string;
  mothersName: string;
  mothersProfession: string;
}